Add support for rich kafka config
