+++ metadata-ingestion/src/gometa/configuration/__init__.py
@@ -1,3 +1,1 @@
-DEFAULT_KAFKA_TOPIC="MetadataChangeEvent_v4"
 from .common import ConfigModel, DynamicTypedConfig, ConfigurationMechanism
-from .kafka import KafkaConnectionConfig
